requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
altair==4.1.0             # via streamlit
appdirs==1.4.4            # via pyppeteer
argon2-cffi==20.1.0       # via notebook
astor==0.8.1              # via streamlit
async-generator==1.10     # via nbclient
attrs==20.2.0             # via jsonschema
backcall==0.2.0           # via ipython
base58==2.0.1             # via streamlit
beautifulsoup4==4.9.3     # via -r requirements.in, bs4
bleach==3.2.1             # via nbconvert
blinker==1.4              # via streamlit
boto3==1.15.18            # via streamlit
botocore==1.18.18         # via boto3, s3transfer, streamlit
bs4==0.0.1                # via requests-html
cachetools==4.1.1         # via streamlit
certifi==2020.6.20        # via requests
cffi==1.14.3              # via argon2-cffi
chardet==3.0.4            # via requests
click==7.1.2              # via pip-tools, streamlit
cssselect==1.1.0          # via pyquery
decorator==4.4.2          # via ipython, validators
defusedxml==0.6.0         # via nbconvert
entrypoints==0.3          # via altair, nbconvert
enum-compat==0.0.3        # via streamlit
fake-useragent==0.1.11    # via requests-html
gitdb==4.0.5              # via gitpython
gitpython==3.1.9          # via streamlit
idna==2.10                # via requests
importlib-metadata==3.2.0  # via jsonschema
ipykernel==5.3.4          # via ipywidgets, notebook, pydeck
ipython-genutils==0.2.0   # via nbformat, notebook, traitlets
ipython==7.18.1           # via ipykernel, ipywidgets
ipywidgets==7.5.1         # via pydeck
jedi==0.17.2              # via ipython
jinja2==2.11.2            # via altair, nbconvert, notebook, pydeck
jmespath==0.10.0          # via boto3, botocore
jsonschema==3.2.0         # via altair, nbformat
jupyter-client==6.1.7     # via ipykernel, nbclient, notebook
jupyter-core==4.6.3       # via jupyter-client, nbconvert, nbformat, notebook
jupyterlab-pygments==0.1.2  # via nbconvert
llvmlite==0.34.0          # via numba
lxml==4.6.0               # via pyquery
markupsafe==1.1.1         # via jinja2
mistune==0.8.4            # via nbconvert
mpmath==1.1.0             # via sympy
nbclient==0.5.1           # via nbconvert
nbconvert==6.0.7          # via notebook
nbformat==5.0.8           # via ipywidgets, nbclient, nbconvert, notebook
nest-asyncio==1.4.1       # via nbclient
notebook==6.1.4           # via widgetsnbextension
numba==0.51.2             # via -r requirements.in
numpy==1.19.2             # via -r requirements.in, altair, numba, pandas, pyarrow, pydeck, scipy, streamlit
packaging==20.4           # via bleach, streamlit
pandas==1.1.3             # via -r requirements.in, altair, streamlit
pandocfilters==1.4.2      # via nbconvert
parse==1.18.0             # via requests-html
parso==0.7.1              # via jedi
pathtools==0.1.2          # via watchdog
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==8.0.0             # via streamlit
pip-tools==5.4.0          # via -r requirements.in
plotly==4.11.0            # via -r requirements.in
prometheus-client==0.8.0  # via notebook
prompt-toolkit==3.0.8     # via ipython
protobuf==3.13.0          # via streamlit
ptyprocess==0.6.0         # via pexpect, terminado
pyarrow==1.0.1            # via streamlit
pycparser==2.20           # via cffi
pydeck==0.4.1             # via streamlit
pyee==7.0.4               # via pyppeteer
pygments==2.7.1           # via ipython, jupyterlab-pygments, nbconvert
pyparsing==2.4.7          # via packaging
pyppeteer==0.2.2          # via requests-html
pyquery==1.4.1            # via requests-html
pyrsistent==0.17.3        # via jsonschema
python-dateutil==2.8.1    # via botocore, jupyter-client, pandas, streamlit
python-dotenv==0.14.0     # via -r requirements.in
pytz==2020.1              # via pandas, tzlocal
pyyaml==5.3.1             # via -r requirements.in
pyzmq==19.0.2             # via jupyter-client, notebook
requests-html==0.10.0     # via -r requirements.in
requests==2.24.0          # via requests-html, streamlit
retrying==1.3.3           # via plotly
s3transfer==0.3.3         # via boto3
scipy==1.5.2              # via -r requirements.in
send2trash==1.5.0         # via notebook
six==1.15.0               # via argon2-cffi, bleach, jsonschema, packaging, pip-tools, plotly, protobuf, python-dateutil, retrying, validators, w3lib
smmap==3.0.4              # via gitdb
soupsieve==2.0.1          # via beautifulsoup4
streamlit==0.69.2         # via -r requirements.in
sympy==1.7.1              # via -r requirements.in
terminado==0.9.1          # via notebook
testpath==0.4.4           # via nbconvert
toml==0.10.1              # via streamlit
toolz==0.11.1             # via altair
tornado==6.0.4            # via ipykernel, jupyter-client, notebook, streamlit, terminado
tqdm==4.50.2              # via pyppeteer
traitlets==5.0.5          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbclient, nbconvert, nbformat, notebook, pydeck
typing-extensions==3.7.4.3  # via importlib-metadata
tzlocal==2.1              # via streamlit
urllib3==1.25.10          # via botocore, pyppeteer, requests
validators==0.18.1        # via streamlit
w3lib==1.22.0             # via requests-html
watchdog==0.10.3          # via streamlit
wcwidth==0.2.5            # via prompt-toolkit
webencodings==0.5.1       # via bleach
websockets==8.1           # via pyppeteer
widgetsnbextension==3.5.1  # via ipywidgets
yapf==0.30.0              # via -r requirements.in
zipp==3.4.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
